% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_match.R
\name{check_match}
\alias{check_match}
\title{Test whether two strings match with a zero-shot LLM prompt.}
\usage{
check_match(name1, name2)
}
\arguments{
\item{name1}{A string}

\item{name2}{A string}
}
\value{
"Yes" if the two strings match, "No" otherwise.
}
\description{
Test whether two strings match with a zero-shot LLM prompt.
}
\examples{
check_match('UPS', 'United Parcel Service')
check_match('UPS', 'United States Postal Service')
check_match('USPS', 'United Parcel Service')
check_match('USPS', 'United States Postal Service')
check_match('USPS', 'Post Office')
}
