% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{dot}
\alias{dot}
\title{Compute the dot product between two vectors}
\usage{
dot(vec1, vec2)
}
\arguments{
\item{vec1}{A numeric vector}

\item{vec2}{Another numeric vector}
}
\value{
A numeric
}
\description{
Compute the dot product between two vectors
}
\examples{
dot(c(0,1), c(1,0))
}
