% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_similarity_matrix.R
\name{get_similarity_matrix}
\alias{get_similarity_matrix}
\title{Create matrix of embedding similarities}
\usage{
get_similarity_matrix(embeddings, strings_A, strings_B)
}
\arguments{
\item{embeddings}{A matrix of text embeddings}

\item{strings_A}{A list of strings in Dataset A}

\item{strings_B}{A list of strings in Dataset B}
}
\value{
A similarity matrix
}
\description{
Create a matrix of pairwise similarities between each string in \code{strings_A} and \code{strings_B}.
}
\examples{
get_embeddings(c('UPS', 'USPS', 'Postal Service')) |>
  get_similarity_matrix(c('UPS', 'USPS'), 'Postal Service')
}
