% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_label.R
\name{hand_label}
\alias{hand_label}
\title{Hand Label A Dataset}
\usage{
hand_label(df)
}
\arguments{
\item{df}{A dataframe with a column called \code{A} and a column called \code{B}}
}
\value{
A labeled dataframe (\code{match} column)
}
\description{
This function prompts the user to manually label a set of name pairs through the \code{R} console.
}
