% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_api_key.R
\name{openai_api_key}
\alias{openai_api_key}
\title{Install an OPENAI API KEY in Your \code{.Renviron} File for Repeated Use}
\usage{
openai_api_key(key, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{key}{The API key provided to you from OpenAI formated in quotes. A key can be acquired at \url{https://platform.openai.com/api-keys}}

\item{overwrite}{If this is set to TRUE, it will overwrite an existing OPENAI_API_KEY that you already have in your \code{.Renviron} file.}

\item{install}{if TRUE, will install the key in your \code{.Renviron} file for use in future sessions.  Defaults to FALSE.}
}
\description{
This function will add your OpenAI API key to your \code{.Renviron} file so it can be called securely without being stored
in your code. After you have installed your key, it can be called any time by typing \code{Sys.getenv("OPENAI_API_KEY")} and will be
automatically called in package functions. If you do not have an \code{.Renviron} file, the function will create one for you.
If you already have an \code{.Renviron} file, the function will append the key to your existing file, while making a backup of your
original file for disaster recovery purposes.
}
\examples{

\dontrun{
openai_api_key("111111abc", install = TRUE)
# First time, reload your environment so you can use the key without restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("OPENAI_API_KEY")
}

\dontrun{
# If you need to overwrite an existing key:
openai_api_key("111111abc", overwrite = TRUE, install = TRUE)
# First time, reload your environment so you can use the key without restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("OPENAI_API_KEY")
}
}
