% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openrouter_api_key.R
\name{openrouter_api_key}
\alias{openrouter_api_key}
\title{Install an OPENROUTER API KEY in Your \code{.Renviron} File for Repeated Use}
\usage{
openrouter_api_key(key, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{key}{The API key provided to you from OpenRouter formatted in quotes. A key can be acquired at \url{https://openrouter.ai/keys}}

\item{overwrite}{If this is set to TRUE, it will overwrite an existing OPENROUTER_API_KEY that you already have in your \code{.Renviron} file.}

\item{install}{if TRUE, will install the key in your \code{.Renviron} file for use in future sessions.  Defaults to FALSE.}
}
\value{
No return value, called for side effects.
}
\description{
This function will add your OpenRouter API key to your \code{.Renviron} file so it can be called securely without being stored
in your code. After you have installed your key, it can be called any time by typing \code{Sys.getenv("OPENROUTER_API_KEY")} and will be
automatically called in package functions. If you do not have an \code{.Renviron} file, the function will create one for you.
If you already have an \code{.Renviron} file, the function will append the key to your existing file, while making a backup of your
original file for disaster recovery purposes.
}
\examples{

\dontrun{
openrouter_api_key("sk-or-v1-abc", install = TRUE)
# First time, reload your environment so you can use the key without restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("OPENROUTER_API_KEY")
}

\dontrun{
# If you need to overwrite an existing key:
openrouter_api_key("sk-or-v1-abc", overwrite = TRUE, install = TRUE)
# First time, reload your environment so you can use the key without restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("OPENROUTER_API_KEY")
}
}
