% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_match.R
\name{check_match}
\alias{check_match}
\title{Test whether two strings match with a zero-shot LLM prompt.}
\usage{
check_match(string1, string2)
}
\arguments{
\item{string1}{A string or vector of strings}

\item{string2}{A string or vector of strings}
}
\value{
"Yes" if the pair of strings match, "No" otherwise.
}
\description{
Test whether two strings match with a zero-shot LLM prompt.
}
