% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_match.R
\name{check_match}
\alias{check_match}
\title{Test whether two strings match with an LLM prompt.}
\usage{
check_match(
  string1,
  string2,
  batch_size = 50,
  model = "gpt-3.5-turbo-instruct",
  few_shot_examples = NULL
)
}
\arguments{
\item{string1}{A string or vector of strings}

\item{string2}{A string or vector of strings}

\item{batch_size}{The number of string pairs to include in each prompt}

\item{model}{Which OpenAI model to prompt; defaults to 'gpt-3.5-turbo-instruct'}

\item{few_shot_examples}{A dataframe with few-shot examples for prompt; must include columns \code{A}, \code{B}, and \code{match}}
}
\value{
A vector the same length as \code{string1} and \code{string2}. "Yes" if the pair of strings match, "No" otherwise.
}
\description{
Test whether two strings match with an LLM prompt.
}
\examples{
check_match('UPS', 'United Parcel Service')
check_match('UPS', 'United States Postal Service')
check_match(c('USPS', 'USPS', 'USPS'),
            c('Post Office', 'United Parcel', 'US Postal Service'))
}
