% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzylink.R
\name{fuzzylink}
\alias{fuzzylink}
\title{Probabilistic Record Linkage Using Pretrained Text Embeddings}
\usage{
fuzzylink(
  dfA,
  dfB,
  by,
  blocking.variables = NULL,
  verbose = TRUE,
  record_type = "entity",
  openai_api_key = NULL,
  max_validations = 10000
)
}
\arguments{
\item{dfA, dfB}{A pair of data frames or data frame extensions (e.g. tibbles)}

\item{by}{A character denoting the name of the variable to use for fuzzy matching}

\item{blocking.variables}{A character vector of variables that must match exactly in order to match two records}

\item{verbose}{TRUE to print progress updates, FALSE for no output}

\item{record_type}{A character describing what type of entity the \code{by} variable represents. Should be a singular noun (e.g. "person", "organization", "interest group", "city").}

\item{openai_api_key}{Your OpenAI API key. By default, looks for a system environment variable called "OPENAI_API_KEY" (recommended option). Otherwise, it will prompt you to enter the API key as an argument.}

\item{max_validations}{The maximum number of LLM prompts to submit during the validation stage; defaults to 10,000}
}
\value{
A dataframe with all rows of \code{dfA} joined with any matches from \code{dfB}
}
\description{
Probabilistic Record Linkage Using Pretrained Text Embeddings
}
\examples{
dfA <- data.frame(state.x77)
dfA$name <- rownames(dfA)
dfB <- data.frame(name = state.abb, state.division)
df <- fuzzylink(dfA, dfB, by = 'name', record_type = 'state')
}
