% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_embeddings.R
\name{get_embeddings}
\alias{get_embeddings}
\title{Get pretrained GPT-3 text embeddings}
\usage{
get_embeddings(text, model = "text-embedding-3-large")
}
\arguments{
\item{text}{A character vector}

\item{model}{Which variant of the GPT-3 embedding model to use. Defaults to 'text-embedding-3-large'.}
}
\value{
A matrix of embedding vectors (one per row).
}
\description{
Get pretrained text embeddings from the OpenAI API. Automatically batches requests to handle rate limits.
}
\examples{
embeddings <- get_embeddings(c('dog', 'cat', 'canine', 'feline'))
embeddings['dog',] |> dot(embeddings['canine',])
embeddings['dog',] |> dot(embeddings['feline',])
}
