% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_embeddings.R
\name{get_embeddings}
\alias{get_embeddings}
\title{Get pre-trained text embeddings from OpenAI API}
\usage{
get_embeddings(text, model = "text-embedding-ada-002")
}
\arguments{
\item{text}{A character vector}

\item{model}{Which variant of the GPT-3 embedding model to use. Defaults to 'text-embedding-ada-002'.}
}
\value{
A named list of embedding vectors, of the same length as the text input.
}
\description{
Get pre-trained text embeddings from OpenAI API
}
\examples{
embeddings <- get_embeddings(c('dog', 'cat', 'canine', 'feline'))
dot(embeddings[['dog']], embeddings[['canine']])
dot(embeddings[['dog']], embeddings[['feline']])
}
