% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_set.R
\name{get_training_set}
\alias{get_training_set}
\title{Create a training set}
\usage{
get_training_set(
  sim,
  num_bins = 50,
  samples_per_bin = 10,
  n = 500,
  record_type = "entity"
)
}
\arguments{
\item{sim}{A matrix of similarity scores}

\item{num_bins}{Number of bins to split similarity scores for stratified random sampling (defaults to 50)}

\item{samples_per_bin}{Number of string pairs to sample from each bin (defaults to 5)}

\item{n}{Sample size for the training dataset}

\item{record_type}{A character describing what type of entity the rows and columns of \code{sim} represent. Should be a singular noun (e.g. "person", "organization", "interest group", "city").}
}
\value{
A dataset with string pairs \code{A} and \code{B}, along with a \code{match} column indicating whether they match.
}
\description{
Creates a training set from a list of similarity matrices and labels it using a zero-shot GPT prompt.
}
